windows_requirements.txt
anyio==3.4.0
argcomplete==1.12.3
asgiref==3.4.1
backcall==0.2.0
click==8.0.3
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
fastapi==0.70.1
gunicorn==20.1.0
h11==0.12.0
idna==3.3
importlib-metadata==4.10.0
ipykernel==6.6.0
ipython==7.30.1
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.21.5
opencv-python-headless==4.5.4.60
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.24
pydantic==1.8.2
Pygments==2.10.0
python-dateutil==2.8.2
python-multipart==0.0.5
pyzmq==22.3.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tflite-runtime==2.5.0
tornado==6.1
traitlets==5.1.1
typing-extensions==4.0.1
uvicorn==0.16.0
wcwidth==0.2.5
zipp==3.6.0